Move unused imports rule into its own module
